+++ src/errors/MissingSubscriptionTopicsError.ts
@@ -1,6 +1,6 @@
 export class MissingSubscriptionTopicsError extends Error {
   constructor(target: Function, methodName: string) {
-    super(`${target}#${methodName} subscription has no provided topics!`);
+    super(`${target.name}#${methodName} subscription has no provided topics!`);
 
     Object.setPrototypeOf(this, new.target.prototype);
   }