+++ tests/functional/subscriptions.ts
@@ -535,7 +535,7 @@ describe("Subscriptions", () => {
 
     it("should throw error while passing empty topics array to Subscription", async () => {
       getMetadataStorage().clear();
-      expect.assertions(4);
+      expect.assertions(5);
       try {
         @ObjectType()
         class SampleObject {
@@ -569,6 +569,7 @@ describe("Subscriptions", () => {
         expect(err).toBeInstanceOf(MissingSubscriptionTopicsError);
         expect(err.message).toContain("SampleResolver");
         expect(err.message).toContain("sampleSubscription");
+        expect(err.message).not.toContain("class SampleResolver");
       }
     });